haruuback/requirements.txt
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.12
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-model-utils==4.0.0
django-rest-auth==0.9.5
django-silk==4.0.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.1
gprof2dot==2019.11.30
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
oauthlib==3.1.0
pycodestyle==2.5.0
Pygments==2.6.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
wincertstore==0.2